Tweak method signature to match changes done in trunk
git-svn-id: https://svn.apache.org/repos/asf/poi/branches/ooxml@634991 13f79535-47bb-0310-9956-ffa450edef68

+++ src/ooxml/interfaces-jdk15/org/apache/poi/ss/usermodel/Row.java
@@ -79,14 +79,12 @@ public interface Row extends Iterable {
      * @param cellnum  0 based column number
      * @return HSSFCell representing that column or null if undefined.
      */
-
-    Cell getCell(short cellnum);
+    Cell getCell(int cellnum);
 
     /**
      * get the number of the first cell contained in this row.
      * @return short representing the first logical cell in the row, or -1 if the row does not contain any cells.
      */
-
     short getFirstCellNum();
 
     /**
